Fixed feature detector as per PR change request
